+++ packages/experimental/NativeButton/ios/ButtonViewManager.swift
@@ -1,7 +1,7 @@
 import Foundation
 import FluentUI
 
-@objc(MSFButtonViewManager)
+@objc(FRNButtonViewManager)
 class ButtonViewManager: RCTViewManager {
 
 	override func view() -> UIView! {