+++ packages/experimental/NativeDatePicker/ios/FRNDatePickerManager.m
@@ -5,7 +5,7 @@
 
 @import FluentUI;
 
-static NSDictionary<NSString *, NSDictionary<NSString *, NSNumber *> *> *MSFDatePickerGetEnumMaps() {
+static NSDictionary<NSString *, NSDictionary<NSString *, NSNumber *> *> *FRNDatePickerGetEnumMaps() {
     return @{
         @"MSFDateTimePickerMode": @{
                 @"date": @(MSFDateTimePickerModeDate),
@@ -24,28 +24,28 @@ static NSDictionary<NSString *, NSDictionary<NSString *, NSNumber *> *> *MSFDate
     };
 }
 
-@implementation RCTConvert (MSFDatePickerAdditions)
+@implementation RCTConvert (FRNDatePickerAdditions)
 
 RCT_ENUM_CONVERTER(MSFDateTimePickerMode,
-                   MSFDatePickerGetEnumMaps()[@"MSFDateTimePickerMode"],
+				   FRNDatePickerGetEnumMaps()[@"MSFDateTimePickerMode"],
                    MSFDateTimePickerModeDate,
                    integerValue)
 RCT_ENUM_CONVERTER(MSFDateTimePickerDatePickerType,
-                   MSFDatePickerGetEnumMaps()[@"MSFDateTimePickerDatePickerType"],
+				   FRNDatePickerGetEnumMaps()[@"MSFDateTimePickerDatePickerType"],
                    MSFDateTimePickerDatePickerTypeCalendar,
                    integerValue)
 RCT_ENUM_CONVERTER(MSFDateTimePickerDateRangePresentation,
-                   MSFDatePickerGetEnumMaps()[@"MSFDateTimePickerDateRangePresentation"],
+				   FRNDatePickerGetEnumMaps()[@"MSFDateTimePickerDateRangePresentation"],
                    MSFDateTimePickerDateRangePresentationPaged,
                    integerValue)
 
 @end
 
-@interface RCT_EXTERN_MODULE(MSFDatePickerManager, NSObject)
+@interface RCT_EXTERN_MODULE(FRNDatePickerManager, NSObject)
 
 - (NSDictionary *)constantsToExport
 {
-    return MSFDatePickerGetEnumMaps();
+    return FRNDatePickerGetEnumMaps();
 }
 
 RCT_EXTERN_METHOD(presentWithMode:(MSFDateTimePickerMode)mode